% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetic.R, R/colors.R
\docType{package}
\name{aesthetic}
\alias{aesthetic}
\alias{aesthetic-package}
\alias{aesthetic}
\title{aesthetic}
\usage{
aesthetic(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{vaporwave}, \code{cool},  \code{crystalpepsi},
\code{mallsoft}, \code{jazzcup},  \code{sunset}, \code{macplus},
\code{seapunk}}

\item{n}{Number of colors desired.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
list of vaporwave-style palettes

The color palettes are taken from https://github.com/mackenziedg/vapeplot
}
\details{
list of vaporwave-style palettes
}
\examples{
wes_palette("vaporwave")
wes_palette("seapunk", 3, "continuous")
}
\keyword{colors}
