% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 8.}
\usage{
palettes
}
\description{
Use \code{\link{aesthetic}} to construct palettes of desired length.
}
\keyword{datasets}
